created sub functions
