+++ app/Http/Controllers/CalendarController.php
@@ -1,0 +1,101 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class CalendarController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+    	return view('pages.athlete-calendar');
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+    	return view('pages.athlete-calendar');
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+    	
+        // Determine what form is being submitted.
+    	if ($request->has('submit_strength')) {
+    		//handle form1
+    		echo "<script>(function(){alert('strength');})();</script>";die();
+    	} else if ($request->has('submit_endurance')) {
+    		//handle form2
+    		print_r("endurance");die();
+    	}else if ($request->has('submit_flexability')) {
+    		//handle form2
+    		print_r("flexability");die();
+    	}else if ($request->has('submit_balance')) {
+    		//handle form2
+    		print_r("balance");die();
+    	}else if ($request->has('submit_notes')) {
+    		//handle form2
+    		print_r("notes");die();
+    	}
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        //
+    }
+}
