+++ routes/web.php
@@ -21,9 +21,9 @@ Route::get('/home', 'HomeController@index')->name('home');
 Route::get('athletes', 'PagesController@getAthletes');
 Route::get('coachcalendar', 'PagesController@getCoachCalendar');
 Route::get('notifications', 'PagesController@getNotifications');
-Route::get('athletecalendar', 'PagesController@getAthleteCalendar');
 Route::get('goals', 'PagesController@getGoals');
 Route::get('schedule', 'PagesController@getSchedules');
 Route::post('search_code', 'FindTrainerController@search_code');
 Route::resource('findtrainer', 'FindTrainerController');
+Route::resource('athletecalendar', 'CalendarController');
 Route::resource('bios', 'BioController');
